started sprint 4 - creared extended user service and updated api

+++ src/main/java/iob/logic/users/UsersServiceJPA.java
@@ -3,22 +3,24 @@ package iob.logic.users;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 import javax.annotation.PostConstruct;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Sort.Direction;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import iob.data.UserEntity;
-import iob.logic.instances.UserId;
 
 @Service
-public class UsersServiceJPA implements UsersService {
+public class UsersServiceJPA implements ExtendedUsersService {
 	private String configurableUser;
 	private UsersCrud userCrud;
-	private userConverter userConverter;
+	private UserConverter userConverter;
 
 	@Autowired
-	public UsersServiceJPA(UsersCrud messageCrud, userConverter messageConverter) {
+	public UsersServiceJPA(UsersCrud messageCrud, UserConverter messageConverter) {
 		this.userCrud = messageCrud;
 		this.userConverter = messageConverter;
 	}
@@ -58,20 +60,17 @@ public class UsersServiceJPA implements UsersService {
 	}
 
 	@Override
-	@Transactional(readOnly = true)
 	public List<UserBoundary> getAllUsers() {
-		Iterable<UserEntity> iter = this.userCrud.findAll();
-		List<UserBoundary> rv = new ArrayList<>();
-		for (UserEntity usr : iter) {
-			rv.add(this.userConverter.toBoundary(usr));
-		}
-		return rv;
+		throw new RuntimeException("deprecated method - use getAllUsers with paginayion instead");
 	}
 
 	@Override
 	@Transactional
-	public void deleteAllUsers() {
-		this.userCrud.deleteAll();	
+	public void deleteAllUsers(String userDomain, String userEmail) {
+		UserEntity entity = getUserEntityById(this.userConverter.toEntity(new UserId(userDomain, userEmail)));
+		if(entity.getRole().toString().equals(UserRoleLogic.ADMIN.toString().toLowerCase())) {
+			this.userCrud.deleteAll();	
+		}
 	}
 	
 	private UserEntity getUserEntityById(String id) {
@@ -84,5 +83,30 @@ public class UsersServiceJPA implements UsersService {
 			throw new UserNotFoundException("could not find message by id: " + id);
 		}
 	}
-	
+
+	@Override
+	@Transactional(readOnly = true)
+	public List<UserBoundary> getAllUsers(int size, int page, String domain, String email) {
+		UserEntity entity = getUserEntityById(this.userConverter.toEntity(new UserId(domain, email)));
+		if(entity.getRole().toString().equals(UserRoleLogic.ADMIN.toString().toLowerCase())) {
+			return this.userCrud
+					.findAll(PageRequest.of(page, size, Direction.ASC, "userId"))
+					.stream() // Stream<MessageEntity>
+					.map(this.userConverter::toBoundary) // Stream<Message>
+					.collect(Collectors.toList()); // List<Message>
+		}
+		else {
+			return new ArrayList<>();
+		}
+	}
+
+	@Override
+	@Transactional(readOnly = true)
+	public List<UserBoundary> getUsersByVersion(int version, int size, int page) {
+		return this.userCrud
+				.findAllByVersion(version, PageRequest.of(page, size, Direction.ASC, "userId"))
+				.stream()
+				.map(this.userConverter::toBoundary)
+				.collect(Collectors.toList());
+	}
 }
